+++ R/test-that.R
@@ -46,7 +46,7 @@ test_that <- function(desc, code) {
     }
   }
 
-  test_code(code, env = parent.frame())
+  test_code(code, parent.frame())
 }
 
 # Access error fields with `[[` rather than `$` because the
@@ -63,6 +63,14 @@ test_code <- function(code, env, reporter = NULL, skip_on_empty = TRUE) {
     reporter$start_test(context = reporter$.context, test = test)
     withr::defer(reporter$end_test(context = reporter$.context, test = test))
   }
+
+  # Want to skip if the test (and its subtests) have no expectations
+  # if (the$top_level_test) {
+  #   the$test_expectations <- 0
+  #   the$top_level_test <- FALSE
+  #   withr::defer(the$top_level_test <- TRUE)
+  # }
+  starting_expectations <- the$test_expectations
 
   ok <- TRUE
 
@@ -88,13 +96,8 @@ test_code <- function(code, env, reporter = NULL, skip_on_empty = TRUE) {
   expressions_opt <- getOption("expressions")
   expressions_opt_new <- min(expressions_opt + 500L, 500000L)
 
-  # If no handlers are called we skip: BDD (`describe()`) tests are often
-  # nested and the top level might not contain any expectations, so we need
-  # some way to disable
-  handled <- !skip_on_empty
-
   handle_error <- function(e) {
-    handled <<- TRUE
+    the$test_expectations <- the$test_expectations + 1L
 
     # Increase option(expressions) to handle errors here if possible, even in
     # case of a stack overflow. This is important for the DebugReporter.
@@ -109,11 +112,11 @@ test_code <- function(code, env, reporter = NULL, skip_on_empty = TRUE) {
     invokeRestart("end_test")
   }
   handle_fatal <- function(e) {
-    handled <<- TRUE
+    the$test_expectations <- the$test_expectations + 1L
     register_expectation(e, 0)
   }
   handle_expectation <- function(e) {
-    handled <<- TRUE
+    the$test_expectations <- the$test_expectations + 1L
     register_expectation(e, 7)
     # Don't bubble up to any other handlers
     invokeRestart("continue_test")
@@ -143,7 +146,7 @@ test_code <- function(code, env, reporter = NULL, skip_on_empty = TRUE) {
     }
   }
   handle_skip <- function(e) {
-    handled <<- TRUE
+    the$test_expectations <- the$test_expectations + 1L
 
     debug_end <- if (inherits(e, "skip_empty")) -1 else 2
     register_expectation(e, debug_end)
@@ -168,7 +171,8 @@ test_code <- function(code, env, reporter = NULL, skip_on_empty = TRUE) {
       withCallingHandlers(
         {
           eval(code, test_env)
-          if (!handled && !is.null(test)) {
+          has_expectations <- the$test_expectations > starting_expectations
+          if (!has_expectations && skip_on_empty) {
             skip_empty()
           }
         },