+++ R/testthat-package.R
@@ -19,7 +19,8 @@
 NULL
 
 the <- new.env(parent = emptyenv())
-the$description <- character()
+the$top_level_test <- TRUE
+the$test_expectations <- 0
 
 # The following block is used by usethis to automatically manage
 # roxygen namespace tags. Modify with care!